backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
intersystems-irispython
sentence-transformers
torch
numpy
python-dotenv==1.0.0
